requirements.txt
# Django Core
Django==6.0.1
asgiref==3.11.0
sqlparse==0.5.5

# Authentication
django-allauth==65.14.0

# Database
dj-database-url==3.1.2
psycopg2-binary==2.9.11

# Production Server
gunicorn==25.1.0
whitenoise==6.12.0

# Environment Variables
python-dotenv==1.0.0

# Image Processing
pillow==12.1.0

# Security
cryptography==46.0.3
PyJWT==2.10.1

# Utilities
requests==2.32.5
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3
six==1.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3
